Edit server.js to use .env
