--- backend/server.js
@@ -1,21 +1,0 @@
-const express = require('express');
-const mongoose = require('mongoose');
-
-const app = express();
-
-// DB configuration
-const db = require('./config/keys').mongoURI;
-
-// Connect to MongoDB
-mongoose
-    .connect(
-        db,
-        { useNewUrlParser: true })
-    .then(() => console.log('MongoDB Connected'))
-    .catch(err => console.log(err));
-// Basic server configuration
-app.get('/', (req, res) => res.send("Hello World"));
-
-const port = process.env.PORT || 5000;
-
-app.listen(port, () => console.log(`Server running on port ${port}`));
